Update cool config mock
