+++ packages/kbn-reporting/export_types/csv/csv_searchsource.test.ts
@@ -23,6 +23,7 @@ import { Writable } from 'stream';
 import { CancellationToken } from '@kbn/reporting-common';
 import { discoverPluginMock } from '@kbn/discover-plugin/server/mocks';
 import { dataPluginMock } from '@kbn/data-plugin/server/mocks';
+import { createMockConfigSchema } from '@kbn/reporting-mocks-server';
 import { createMockScreenshottingStart } from '@kbn/screenshotting-plugin/server/mock';
 
 import { CsvSearchSourceExportType } from '.';
