+++ packages/kbn-reporting/export_types/png/png_v2.test.ts
@@ -13,6 +13,7 @@ import { coreMock, elasticsearchServiceMock, loggingSystemMock } from '@kbn/core
 import { CancellationToken } from '@kbn/reporting-common';
 import type { LocatorParams } from '@kbn/reporting-common/types';
 import type { TaskPayloadPNGV2 } from '@kbn/reporting-export-types-png-common';
+import { createMockConfigSchema } from '@kbn/reporting-mocks-server';
 import { cryptoFactory, generatePngObservable } from '@kbn/reporting-server';
 import type { ScreenshottingStart } from '@kbn/screenshotting-plugin/server';
