+++ x-pack/plugins/reporting/server/test_helpers/create_mock_reportingplugin.ts
@@ -8,6 +8,8 @@
 jest.mock('../routes');
 jest.mock('../usage');
 
+import { BehaviorSubject } from 'rxjs';
+
 import {
   coreMock,
   docLinksServiceMock,
@@ -21,12 +23,12 @@ import { featuresPluginMock } from '@kbn/features-plugin/server/mocks';
 import { FieldFormatsRegistry } from '@kbn/field-formats-plugin/common';
 import { fieldFormatsMock } from '@kbn/field-formats-plugin/common/mocks';
 import { licensingMock } from '@kbn/licensing-plugin/server/mocks';
+import { createMockConfigSchema } from '@kbn/reporting-mocks-server';
 import { setFieldFormats } from '@kbn/reporting-server';
 import { createMockScreenshottingStart } from '@kbn/screenshotting-plugin/server/mock';
 import { securityMock } from '@kbn/security-plugin/server/mocks';
 import { taskManagerMock } from '@kbn/task-manager-plugin/server/mocks';
-import { BehaviorSubject } from 'rxjs';
-import { DeepPartial } from 'utility-types';
+
 import { ReportingCore } from '..';
 import { ReportingConfigType } from '../config';
 import { ReportingInternalSetup, ReportingInternalStart } from '../core';
@@ -81,44 +83,6 @@ export const createMockPluginStart = async (
     screenshotting: createMockScreenshottingStart(),
     ...startMock,
   };
-};
-
-export const createMockConfigSchema = (
-  overrides: DeepPartial<ReportingConfigType> = {}
-): ReportingConfigType => {
-  // deeply merge the defaults and the provided partial schema
-  return {
-    index: '.reporting',
-    encryptionKey: 'cool-encryption-key-where-did-you-find-it',
-    ...overrides,
-    kibanaServer: {
-      hostname: 'localhost',
-      ...overrides.kibanaServer,
-    },
-    queue: {
-      indexInterval: 'week',
-      pollEnabled: true,
-      pollInterval: 3000,
-      timeout: 120000,
-      ...overrides.queue,
-    },
-    csv: {
-      scroll: { size: 500, duration: '30s' },
-      ...overrides.csv,
-    },
-    roles: {
-      enabled: false,
-      ...overrides.roles,
-    },
-    capture: { maxAttempts: 1 },
-    export_types: {
-      pdf: { enabled: true },
-      png: { enabled: true },
-      csv: { enabled: true },
-      ...overrides.export_types,
-    },
-    statefulSettings: { enabled: true },
-  } as ReportingConfigType;
 };
 
 export const createMockReportingCore = async (
